feat(vs2022_third): C4244 fix remaining double→float conversions in RecipeStressStageProgramView.cpp — 1 file
Fix 14 remaining C4244 warnings (double→float narrowing conversions) by adding
static_cast<float>() at assignment sites.

No behavior changes, only explicit casts to match target types.

+++ SRC/NanoSpec/RecipeStressStageProgramView.cpp
@@ -2615,7 +2615,7 @@ BOOL CRecipeStressStageProgramView::CheckMeasurementLineSectionData()
 // 2009.09.10 bagus stress èCê≥ --{--
 //	m_iNumScans = NumOfSections * PIN_INTERVAL;
 //	m_StageProgInfoHdr.wNumScans = m_iNumScans;
-	m_StageProgInfoHdr.wNumScans = static_cast<long>(NumOfSections * PIN_INTERVAL);
+	m_StageProgInfoHdr.wNumScans = static_cast<float>(static_cast<long>(NumOfSections * PIN_INTERVAL));
 	sprintf_s(szBuff, _countof(szBuff), "%d", m_StageProgInfoHdr.wNumScans);
 	SetDlgItemText(IDC_STAGE_PROGRAM_POINT_NUMBER, szBuff);
